Refactor SplitClient to use pydantic types
