+++ src/pypaystack2/sub_clients/__init__.py
@@ -24,7 +24,6 @@ Modules exported by this package:
     - `verification`: A module containing implementations for interfacing with Paystack's Verification API
 """
 
-from pypaystack2.sub_clients.splits import TransactionSplitClient
 from pypaystack2.sub_clients.subaccounts import SubAccountClient
 from pypaystack2.sub_clients.subscriptions import SubscriptionClient
 
@@ -45,6 +44,7 @@ from pypaystack2.sub_clients.sync_clients.plans import PlanClient
 from pypaystack2.sub_clients.sync_clients.products import ProductClient
 from pypaystack2.sub_clients.sync_clients.refunds import RefundClient
 from pypaystack2.sub_clients.sync_clients.settlements import SettlementClient
+from pypaystack2.sub_clients.sync_clients.splits import TransactionSplitClient
 from pypaystack2.sub_clients.terminals import TerminalClient
 from pypaystack2.sub_clients.transactions import TransactionClient
 from pypaystack2.sub_clients.transfers import TransferClient
