+++ src/pypaystack2/sub_clients/async_clients/settlements.py
@@ -1,10 +1,9 @@
 from http import HTTPMethod
 from typing import Type
 
-from pypaystack2 import Response
 from pypaystack2.base_api_client import BaseAsyncAPIClient
 from pypaystack2.utils import append_query_params
-from pypaystack2.utils.models import PaystackDataModel
+from pypaystack2.utils.models import PaystackDataModel, Response
 from pypaystack2.utils.response_models import Settlement, Transaction
