+++ src/pypaystack2/utils/response_models.py
@@ -377,14 +377,14 @@ class SubAccount(BaseModel):
     primary_contact_name: str | None = None
     primary_contact_email: str | None = None
     primary_contact_phone: str | None = None
-    metadata: dict[str, Any]
-    percentage_charge: int
+    metadata: dict[str, Any] | None = None
+    percentage_charge: int | None = None
     settlement_bank: str
-    bank_id: int
+    bank_id: int | None = None
     account_number: str
     currency: Currency
-    active: int | bool
-    is_verified: bool
+    active: int | bool | None = None
+    is_verified: bool | None = None
     integration: int | None = None
     bank: str | None = None
     managed_by_integration: int | None = None
